project_ai_academy/requirements.txt
google-genai
gspread
google-auth-oauthlib
google-auth
google-api-python-client
python-dotenv
Pillow
moviepy
numpy
